Backend/requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
mysqlclient
requests
